requirements.txt
Flask==2.2.2
lxml==4.9.1
pandas==1.5.2
PyYAML==6.0
selenium==4.8.0
webdriver_manager==3.8.5
